+++ ng/practices-ng-list-view-source/src/lib/table-view-source-fluent-builder/create-table-view-source-with-persisted-params.ts
@@ -150,10 +150,20 @@ export function extendWithPersistedParams<TTableViewSource extends TableViewSour
                 const latestPersistedParams: IQueryParams | null = await firstValueFromMaybeAsync(
                     config.persistParams.load()
                 );
-                trace('tableViewSourceWithPersistedParams', 'loaded initial query params', latestPersistedParams);
+                trace(
+                    'ListViewSource',
+                    'tableViewSourceWithPersistedParams',
+                    'loading initial query params',
+                    latestPersistedParams,
+                    this
+                );
                 if (latestPersistedParams) {
                     await firstValueFrom(timer(0)); // First timer is needed so that the ui notices the change
                     this.update(latestPersistedParams);
+                    trace('ListViewSource', 'tableViewSourceWithPersistedParams', 'apply loaded query params', {
+                        latestPersistedParams,
+                        current: this.getQueryParams(),
+                    });
                     await firstValueFrom(timer(0)); // Second timer is needed so the queryParams$ subscription below doesn't get triggered unnecessarily.
                 }
 
@@ -166,7 +176,7 @@ export function extendWithPersistedParams<TTableViewSource extends TableViewSour
                         debounceTime(100),
                         switchMap((queryParams) => {
                             if (!isObjDeepEqual(queryParams, latestPersistedParams, { maximumDepth: 3 })) {
-                                trace('tableViewSourceWithPersistedParams', 'persist query params', {
+                                trace('ListViewSource', 'tableViewSourceWithPersistedParams', 'persist query params', {
                                     queryParams,
                                     latestPersistedParams,
                                 });
@@ -184,4 +194,3 @@ export function extendWithPersistedParams<TTableViewSource extends TableViewSour
 
     return this;
 }
-